Add confirm dialog for categoru delete

+++ apps/admin/src/app/categories/categories-list/categories-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoriesService, Category } from '@elobong/products'
-import { MessageService } from 'primeng/api';
+import {ConfirmationService, MessageService} from 'primeng/api';
+
 
 @Component({
   selector: 'admin-categories-list',
@@ -14,7 +15,8 @@ export class CategoriesListComponent implements OnInit {
 
   constructor(
     private categoriesService: CategoriesService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService,
   ) { }
 
   ngOnInit(): void {
@@ -22,17 +24,27 @@ export class CategoriesListComponent implements OnInit {
   }
 
   deleteCategory(categoryId: string) {
-    this.categoriesService.deleteCategory(categoryId).subscribe(res => {
-      this._getCategories();
-      this.messageService.add({
-        severity:'success', summary:'Success', detail: res.message
-      });
-    },
-    (error) => {
-      this.messageService.add({
-        severity:'error', summary:'Error', detail:'Category was not deleted, Please try again later'
-      });
-    })
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to delete this category?',
+      header: 'Delete Category',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.categoriesService.deleteCategory(categoryId).subscribe(res => {
+          this._getCategories();
+          this.messageService.add({
+            severity:'success', summary:'Success', detail: res.message
+          });
+        },
+        (error) => {
+          this.messageService.add({
+            severity:'error', summary:'Error', detail:'Category was not deleted, Please try again later'            
+          });
+          console.error(error);
+        })
+      },
+      // reject: (type:any) => {
+      // }
+  });
   }
 
   private _getCategories(){
